Remove defensive conditional
This conditional is fairly pointless, since we're casting the object to
this protocol in the method signature. Furthermore, since the Class
object doesn't have `+conformsToProtocol:` defined on it (that's defined
on `NSObject`), this conditional throws an unfixable warning about
finding the instance method (defined on the `NSObject` protocol, which
is inherited through `TBSerializable`).